On Mouse Over
Not working yet, but routine seems solid. Need to discover why its not working.

Also made som changes in view (map box): split method was using in strings without necessarily the split marker, so it crashed.

Removed the split methods and need to think in a better strategy of parsing the address in effective layout way